Add added entity tracking + create dedicated root module for filters
