+++ crates/modor/src/entities/mod.rs
@@ -1,4 +1,3 @@
 pub(crate) mod building;
-pub(crate) mod filters;
 pub(crate) mod runner;
 pub(crate) mod traits;
