+++ crates/modor/src/lib.rs
@@ -70,6 +70,7 @@ mod utils;
 mod actions;
 mod app;
 mod entities;
+mod filters;
 mod ranges;
 mod storages;
 mod system_params;
@@ -78,9 +79,13 @@ mod systems;
 pub use actions::*;
 pub use app::*;
 pub use entities::building::*;
-pub use entities::filters::*;
 pub use entities::runner::*;
 pub use entities::traits::*;
+pub use filters::changed::*;
+pub use filters::or::*;
+pub use filters::with::*;
+pub use filters::without::*;
+pub use filters::*;
 pub use ranges::*;
 pub use system_params::entity::*;
 pub use system_params::filters::*;
